requirements.txt
streamlit 
faiss-cpu 
numpy 
sentence-transformers
PyPDF2